+++ src/php/Compiler/Application/GlobalEnvironmentManager.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Phel\Compiler\Application;
 
+use Phel\Compiler\Domain\Analyzer\Environment\GlobalEnvironmentInterface;
 use Phel\Compiler\Infrastructure\GlobalEnvironmentSingleton;
 
 final class GlobalEnvironmentManager
@@ -16,5 +17,25 @@ final class GlobalEnvironmentManager
     public function reset(): void
     {
         GlobalEnvironmentSingleton::reset();
+    }
+
+    public function isInitialized(): bool
+    {
+        return GlobalEnvironmentSingleton::isInitialized();
+    }
+
+    public function getInstance(): GlobalEnvironmentInterface
+    {
+        return GlobalEnvironmentSingleton::getInstance();
+    }
+
+    public function initializeNew(): GlobalEnvironmentInterface
+    {
+        return GlobalEnvironmentSingleton::initializeNew();
+    }
+
+    public function setInstance(GlobalEnvironmentInterface $env): void
+    {
+        GlobalEnvironmentSingleton::setInstance($env);
     }
 }