+++ src/php/Console/ConsoleFactory.php
@@ -7,9 +7,9 @@ namespace Phel\Console;
 use Gacela\Framework\AbstractConfig;
 use Gacela\Framework\AbstractFactory;
 use Phel\Console\Application\ArgvInputSanitizer;
-use Phel\Console\Application\VersionFinder;
 use Phel\Console\Infrastructure\ConsoleBootstrap;
 use Phel\Filesystem\FilesystemFacadeInterface;
+use Phel\Shared\VersionFinder;
 use Symfony\Component\Console\Command\Command;
 
 use function in_array;