+++ tests/php/Unit/Nrepl/Domain/Op/EvalOpTest.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace PhelTest\Unit\Nrepl\Domain\Op;
 
-use Phel\Compiler\Infrastructure\CompileOptions;
 use Phel\Nrepl\Application\Op\EvalOp;
 use Phel\Nrepl\Application\Op\EvalResultResponder;
 use Phel\Nrepl\Domain\Op\OpRequest;
@@ -12,6 +11,7 @@ use Phel\Nrepl\Domain\Session\SessionRegistry;
 use Phel\Printer\PrinterInterface;
 use Phel\Run\Domain\Repl\EvalError;
 use Phel\Run\Domain\Repl\EvalResult;
+use Phel\Shared\CompileOptions;
 use Phel\Shared\Facade\RunFacadeInterface;
 use PHPUnit\Framework\TestCase;